# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 'Sehatly telemedicine API'
  version: 1.0.0
  description: |-
    GENERATED FILE -- do not hand-edit. Run `php artisan sehatly:openapi` to rewrite it and
    `php artisan sehatly:openapi --check` to prove the committed copy is still a fresh export.

    Every path, method, middleware entry and request rule below was read from the running
    application's route table (`Route::getRoutes()`) and from each controller method's injected
    `FormRequest`. Nothing here is transcribed from `routes/api.php`, which is why the file cannot
    drift from the code: a route added, renamed or removed changes it on the next run, and
    `--check` fails until the regenerated file is committed.

    This document describes 77 routes across 68 distinct paths, read from the live route table.

    ## Envelopes

    Every response -- success or failure -- carries the same envelope. `success` is a boolean,
    `message` is a stable human-readable string that does not change with the request, and `errors`
    is present only on failures. A failure with no field-level detail (401, 403, 404, the sanitized
    500) carries `"errors": {}`.

    A list response additionally carries `meta`, as a TOP-LEVEL FOURTH KEY and not nested inside
    `data`. That position is fixed: `App\Support\ApiResponse` appends it after `message` precisely so
    adding pagination cannot renumber the three keys every existing client already reads. A response
...

    ## Validation errors

...

...

...

...

...

...

...

...

...

...
security:
  -
    sanctum: []
paths:
...
  '/api/v1/konsultasi/{id}/resep':
    parameters:
      -
        name: id
        type: integer
        required: true
    post:
      operationId: postApiV1KonsultasiIdResep
      summary: 'Create /api/v1/konsultasi/{id}/resep.'
      tags:
        - konsultasi
      security:
        -
          sanctum: []
      x-route-name: konsultasi.resep.store
      x-laravel-action: App\Http\Controllers\Api\V1\ResepController@store
      x-middleware:
        - 'auth:sanctum'
        - 'tipe:dokter'
        - 'permission:resep.buat'
      x-auth:
        scheme: sanctum
        transport: bearer
      requestBody:
        required: true
        description: 'Validated by `App\Http\Requests\Resep\StoreResepRequest::rules()`, which is read at generation time -- these properties are the live rules, not a transcription.'
        content:
          application/json: { schema: { $ref: '#/components/schemas/StoreResepRequestBody' } }
      responses:
        201:
          description: 'Created. `data` holds the created resource.'
          content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
        422:
          description: 'Validation failed. `message` is the fixed string "The given data was invalid." and `errors` maps each field to an ARRAY of messages -- a field can fail more than one rule, and every message is carried. Keys are the dotted attribute path as submitted (`items.0.obat_id` for an array element).'
          content: { application/json: { schema: { $ref: '#/components/schemas/ValidationErrorEnvelope' } } }
        401:
          description: 'Unauthenticated. No token, an expired token, or a token that was revoked. `message` is the fixed string "Unauthenticated." and `errors` is `{}`.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
        403:
          description: 'Forbidden. The caller is authenticated but holds no grant for this operation (`permission:` middleware), or is not an account type this route allows (`tipe:` middleware). `message` is the fixed string "This action is unauthorized." and `errors` is `{}`.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
        404:
          description: 'Resource not found. Also answers a `{placeholder}` outside the route''s own constraint, such as an unknown `gateway` on the webhook. `message` is the fixed string "Resource not found." -- never a model or table name -- and `errors` is `{}`.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
        500:
          description: 'Internal server error. The body is a fixed sanitized string; the diagnostic detail is kept server-side and never sent to a client.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
  /api/v1/obat:
    get:
      operationId: getApiV1Obat
      summary: 'Read /api/v1/obat.'
      tags:
        - obat
      security:
        -
          sanctum: []
      x-route-name: obat.index
      x-laravel-action: App\Http\Controllers\Api\V1\ResepController@search
      x-middleware:
        - 'auth:sanctum'
        - 'tipe:dokter'
        - 'permission:obat.cari'
      x-auth:
        scheme: sanctum
        transport: bearer
      responses:
        200:
          description: 'Success. `data` holds the requested resource, and `meta` is present when the response is a list.'
          content: { application/json: { schema: { $ref: '#/components/schemas/PaginatedEnvelope' } } }
        422:
          description: 'Validation failed. `message` is the fixed string "The given data was invalid." and `errors` maps each field to an ARRAY of messages -- a field can fail more than one rule, and every message is carried. Keys are the dotted attribute path as submitted (`items.0.obat_id` for an array element).'
          content: { application/json: { schema: { $ref: '#/components/schemas/ValidationErrorEnvelope' } } }
        401:
          description: 'Unauthenticated. No token, an expired token, or a token that was revoked. `message` is the fixed string "Unauthenticated." and `errors` is `{}`.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
        403:
          description: 'Forbidden. The caller is authenticated but holds no grant for this operation (`permission:` middleware), or is not an account type this route allows (`tipe:` middleware). `message` is the fixed string "This action is unauthorized." and `errors` is `{}`.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
        404:
          description: 'Resource not found. Also answers a `{placeholder}` outside the route''s own constraint, such as an unknown `gateway` on the webhook. `message` is the fixed string "Resource not found." -- never a model or table name -- and `errors` is `{}`.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
        500:
          description: 'Internal server error. The body is a fixed sanitized string; the diagnostic detail is kept server-side and never sent to a client.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
  /api/v1/pasien/resep:
    get:
      operationId: getApiV1PasienResep
      summary: 'Read /api/v1/pasien/resep.'
      tags:
        - pasien
      security:
        -
          sanctum: []
      x-route-name: pasien.resep.index
      x-laravel-action: App\Http\Controllers\Api\V1\ResepController@riwayat
      x-middleware:
        - 'auth:sanctum'
        - 'permission:resep.lihat'
      x-auth:
        scheme: sanctum
        transport: bearer
      responses:
        200:
          description: 'Success. `data` holds the requested resource, and `meta` is present when the response is a list.'
          content: { application/json: { schema: { $ref: '#/components/schemas/PaginatedEnvelope' } } }
        422:
          description: 'Validation failed. `message` is the fixed string "The given data was invalid." and `errors` maps each field to an ARRAY of messages -- a field can fail more than one rule, and every message is carried. Keys are the dotted attribute path as submitted (`items.0.obat_id` for an array element).'
          content: { application/json: { schema: { $ref: '#/components/schemas/ValidationErrorEnvelope' } } }
        401:
          description: 'Unauthenticated. No token, an expired token, or a token that was revoked. `message` is the fixed string "Unauthenticated." and `errors` is `{}`.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
        403:
          description: 'Forbidden. The caller is authenticated but holds no grant for this operation (`permission:` middleware), or is not an account type this route allows (`tipe:` middleware). `message` is the fixed string "This action is unauthorized." and `errors` is `{}`.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
        404:
          description: 'Resource not found. Also answers a `{placeholder}` outside the route''s own constraint, such as an unknown `gateway` on the webhook. `message` is the fixed string "Resource not found." -- never a model or table name -- and `errors` is `{}`.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
        500:
          description: 'Internal server error. The body is a fixed sanitized string; the diagnostic detail is kept server-side and never sent to a client.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
  /api/v1/pasien/surat-keterangan:
    get:
      operationId: getApiV1PasienSuratKeterangan
      summary: 'Read /api/v1/pasien/surat-keterangan.'
      tags:
        - pasien
      security:
        -
          sanctum: []
      x-route-name: pasien.surat-keterangan.index
      x-laravel-action: App\Http\Controllers\Api\V1\SuratKeteranganController@daftar
      x-middleware:
        - 'auth:sanctum'
      x-auth:
        scheme: sanctum
        transport: bearer
      responses:
        200:
          description: 'Success. `data` holds the requested resource, and `meta` is present when the response is a list.'
          content: { application/json: { schema: { $ref: '#/components/schemas/PaginatedEnvelope' } } }
        401:
          description: 'Unauthenticated. No token, an expired token, or a token that was revoked. `message` is the fixed string "Unauthenticated." and `errors` is `{}`.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
        403:
          description: 'Forbidden. The caller is authenticated but holds no grant for this operation (`permission:` middleware), or is not an account type this route allows (`tipe:` middleware). `message` is the fixed string "This action is unauthorized." and `errors` is `{}`.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
        404:
          description: 'Resource not found. Also answers a `{placeholder}` outside the route''s own constraint, such as an unknown `gateway` on the webhook. `message` is the fixed string "Resource not found." -- never a model or table name -- and `errors` is `{}`.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
        500:
          description: 'Internal server error. The body is a fixed sanitized string; the diagnostic detail is kept server-side and never sent to a client.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
  /api/v1/resep:
    get:
      operationId: getApiV1Resep
      summary: 'Read /api/v1/resep.'
      tags:
        - resep
      security:
        -
          sanctum: []
      x-route-name: resep.antrean
      x-laravel-action: App\Http\Controllers\Api\V1\ResepController@antrean
      x-middleware:
        - 'auth:sanctum'
        - 'tipe:apoteker'
        - 'permission:resep.verifikasi'
      x-auth:
        scheme: sanctum
        transport: bearer
      responses:
        200:
          description: 'Success. `data` holds the requested resource, and `meta` is present when the response is a list.'
          content: { application/json: { schema: { $ref: '#/components/schemas/PaginatedEnvelope' } } }
        422:
          description: 'Validation failed. `message` is the fixed string "The given data was invalid." and `errors` maps each field to an ARRAY of messages -- a field can fail more than one rule, and every message is carried. Keys are the dotted attribute path as submitted (`items.0.obat_id` for an array element).'
          content: { application/json: { schema: { $ref: '#/components/schemas/ValidationErrorEnvelope' } } }
        401:
          description: 'Unauthenticated. No token, an expired token, or a token that was revoked. `message` is the fixed string "Unauthenticated." and `errors` is `{}`.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
        403:
          description: 'Forbidden. The caller is authenticated but holds no grant for this operation (`permission:` middleware), or is not an account type this route allows (`tipe:` middleware). `message` is the fixed string "This action is unauthorized." and `errors` is `{}`.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
        404:
          description: 'Resource not found. Also answers a `{placeholder}` outside the route''s own constraint, such as an unknown `gateway` on the webhook. `message` is the fixed string "Resource not found." -- never a model or table name -- and `errors` is `{}`.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
        500:
          description: 'Internal server error. The body is a fixed sanitized string; the diagnostic detail is kept server-side and never sent to a client.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
  '/api/v1/resep/{id}':
    parameters:
      -
        name: id
        type: integer
        required: true
    get:
      operationId: getApiV1ResepId
      summary: 'Read /api/v1/resep/{id}.'
      tags:
        - resep
      security:
        -
          sanctum: []
      x-route-name: resep.show
      x-laravel-action: App\Http\Controllers\Api\V1\ResepController@show
      x-middleware:
        - 'auth:sanctum'
        - 'permission:resep.lihat'
      x-auth:
        scheme: sanctum
        transport: bearer
      responses:
        200:
          description: 'Success. `data` holds the requested resource, and `meta` is present when the response is a list.'
          content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
        401:
          description: 'Unauthenticated. No token, an expired token, or a token that was revoked. `message` is the fixed string "Unauthenticated." and `errors` is `{}`.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
        403:
          description: 'Forbidden. The caller is authenticated but holds no grant for this operation (`permission:` middleware), or is not an account type this route allows (`tipe:` middleware). `message` is the fixed string "This action is unauthorized." and `errors` is `{}`.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
        404:
          description: 'Resource not found. Also answers a `{placeholder}` outside the route''s own constraint, such as an unknown `gateway` on the webhook. `message` is the fixed string "Resource not found." -- never a model or table name -- and `errors` is `{}`.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
        500:
          description: 'Internal server error. The body is a fixed sanitized string; the diagnostic detail is kept server-side and never sent to a client.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
  '/api/v1/resep/{id}/cek-interaksi':
    parameters:
      -
        name: id
        type: integer
        required: true
    get:
      operationId: getApiV1ResepIdCekInteraksi
      summary: 'Read /api/v1/resep/{id}/cek-interaksi.'
      tags:
        - resep
      security:
        -
          sanctum: []
      x-route-name: resep.cek-interaksi
      x-laravel-action: App\Http\Controllers\Api\V1\ResepController@cekInteraksi
      x-middleware:
        - 'auth:sanctum'
        - 'permission:resep.lihat'
      x-auth:
        scheme: sanctum
        transport: bearer
      responses:
        200:
          description: 'Success. `data` holds the requested resource, and `meta` is present when the response is a list.'
          content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
        401:
          description: 'Unauthenticated. No token, an expired token, or a token that was revoked. `message` is the fixed string "Unauthenticated." and `errors` is `{}`.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
        403:
          description: 'Forbidden. The caller is authenticated but holds no grant for this operation (`permission:` middleware), or is not an account type this route allows (`tipe:` middleware). `message` is the fixed string "This action is unauthorized." and `errors` is `{}`.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
        404:
          description: 'Resource not found. Also answers a `{placeholder}` outside the route''s own constraint, such as an unknown `gateway` on the webhook. `message` is the fixed string "Resource not found." -- never a model or table name -- and `errors` is `{}`.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
        500:
          description: 'Internal server error. The body is a fixed sanitized string; the diagnostic detail is kept server-side and never sent to a client.'
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
components:
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
      bearerFormat: 'Sanctum personal access token'
      description: 'Send `Authorization: Bearer <token>`. The token is issued by `POST /api/v1/auth/otp/verify` and refreshed by `POST /api/v1/auth/refresh`. Every operation below declares its own `security`, so an empty `security: []` means the route is intentionally anonymous.'
  schemas:
    SuccessEnvelope:
      type: object
      description: '`{"success":true,"data":<data>,"message":<message>}`. Exactly these three keys: there is NO `meta` key, because omission is what makes "this response is not paginated" unambiguous. Key order is `success`, `data`, `message` and is load-bearing.'
      properties:
        success:
          type: boolean
          const: true
        data:
          type: object
          description: "The endpoint's payload. See the operation's notes."
        message:
          type: string
      required:
        - success
        - data
        - message
      additionalProperties: false
...
    ValidationErrorEnvelope:
      type: object
      description: 'The 422 envelope. Identical in shape to `ErrorEnvelope` and deliberately declared as its own component rather than reused, because the thing that makes a 422 hard to consume is `errors` carrying MORE THAN ONE MESSAGE PER FIELD, and a client that treats it as `{field: string}` silently drops every message after the first. `message` is the fixed string "The given data was invalid." -- never `ValidationException::summarize()`, which promotes the first field error and appends "(and N more errors)" and would make `message` data-dependent.'
      properties:
        success:
          type: boolean
          const: false
        message:
          type: string
        errors:
          type: object
          description: 'A field-keyed map of messages. `ApiResponse::error()` casts it to an object so an empty set encodes as `{}` rather than `[]`. Keys are the dotted attribute path as submitted, which is exactly what `ValidationException::errors()` produces.'
          additionalProperties: { type: array, items: { type: string }, minItems: 1, description: 'Every message produced for this field. A field that fails more than one rule carries more than one entry; nothing is flattened into a single string.' }
      required:
        - success
        - message
        - errors
      additionalProperties: false
...